Feature/kafka (#254)
* added sample for mockito

* set the async thread pool as default based on configurations

* added test for async

* more stable rejection policy

* added ADR for centralized asynch

* correcting ADR

* added aop logging the service interface impl this include the execution time

* fixed logging aspect

* added graphql and async configurations

* added graphql and async configurations

* added graphql and async configurations

* formate the code

* running test

* running test

* upgraded to version 3.5.0

* added test for the interceptor

* fixed tests

* fixed tests

* added interceptor on his own package

* added interceptor on his own package

* added metrics test

* added metrics test

* added metrics test for each type

* test signed commit

* test signed commit

* added submodule for user-interface

* rename submodule and project update

* upgraded docker-compose

* Fixed submodule URLs

* added kafka on docker with zookeper

* upgraded to java 25

* upgrade checker

* update readme

* update ADR

* chore: added preconfig grafana graphs

* chore: fixed kafka config issue and upgraded the pipeline

* chore: fixed kafka config issue and upgraded the pipeline

* chore: fixed kafka config issue and upgraded the pipeline

* chore: fixed kafka config issue and upgraded the pipeline

* chore: fixed kafka config issue and upgraded the pipeline

* chore: fixed kafka config issue and upgraded the pipeline

* feat(): setup future documentation

* test

* feat(): countries endpoint and test

* add

* added business key generator to up to 64 instances, not bottom neck on the front end to read the values

* added information to ADR about Null safety, apache pekko and snowflake generation id

* added information to ADR about Opentelemetry

* feat(): fix error cases

* feat(): added get endpoint for customer, to display a list of them

+++ src/main/resources/database/mssql/migration/V0.4__create_machine_registry.sql
@@ -1,6 +1,6 @@
 CREATE TABLE machine_registry
 (
-    machine_id      INT PRIMARY KEY,
+    machine_id      BIGINT PRIMARY KEY,
     instance_id     VARCHAR(255) NOT NULL UNIQUE,
     service_name    VARCHAR(100) NOT NULL,
     allocated_at    DATETIMEOFFSET NOT NULL DEFAULT CURRENT_TIMESTAMP,
